requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# LangGraph agent (supervisor pattern)
langgraph>=0.2.0
langchain-core>=0.3.0

# Optional: LLM-based command parsing
langchain-openai>=0.2.0

# Web server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
